requirements.txt
torch == 0.4.1
torchvision == 0.2.1
numpy >= 1.13.0
matplotlib >= 1.5
Pillow >= 1.1.7
# pip install torch==1.7.1+cu101 torchvision==0.8.2+cu101 torchaudio==0.7.2 -f https://download.pytorch.org/whl/torch_stable.html